Made it possible to run sync/update_translation_fields limited to a single app.
The limiting only works with Django 1.7+. There is a new command line
option "--app" for both commands that takes an app label (last part of
name from INSTALLED_APPS with older apps or whatever is specified in
apps.py with newer). You can also directly pass an AppConfig instance
as a call_command argument, which is handy when executing the commands
as part of a post_migrate handler.