Demangled sync/update_translation_fields from modeltranslation settings.
This lets you register one field with a set of AVAILABLE_LANGUAGES and
a DEFAULT_LANGUAGE, register another one with different settings, and
still have the commands work properly.

Mostly useful for mixing modeltranslation tests with project tests --
project may not include languages enforced by the test settings, leading
to problems when commands are run in a post_migrate handler.