Refactored the main test base to avoid some module reloading.
Reloading the translator module is better avoided at all as it creates a
copy of the registry, making debugging tests harder. A lighter option
could be resetting the registry, but because we're preventing registration
of test models before enforcing test settings this is not necessary.

Reloading modeltranslation settings was leaving references to pre-existing
settings in other modules and that could cause some surprises.

Note that there is still some reloading and some globals black magic.
It would be nice to to get rid of it completely, but we'd need to provide
unpatching of models (such that unregistering would be equivalent to not
registering).