Fix test failures when django directory is not writable
The tests modify the Group model exported by the django.contrib.auth
application and as such, when makemigrations is called, it wants to create
a new file in django/contrib/auth/migrations. This fails when the
directory is not writable, as is the case for a typical system-wide
installation (with a Debian package in my case).

To avoid this problem, we modify the test runner to set the
MIGRATION_MODULES django setting so that the migrations for the auth
application are looked up in the “modeltranslation.tests.auth_migrations”
package and we initialize that package with the migrations found
in the django directory.

Fix #377.